requirements.txt
numpy==1.24.0
Pillow==9.5.0
